requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
gTTS==2.5.4
idna==3.10
mysqlclient==2.2.7
python-dotenv==1.1.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.11.0
reportlab==4.4.4
xhtml2pdf==0.2.17
gunicorn==23.0.0
